requirements.txt
amqp==5.1.1
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
boto3==1.26.84
botocore==1.29.84
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
dj-config-url==0.1.1
Django==4.1.7
django-extensions==3.2.1
django-storages==1.13.2
django-widget-tweaks==1.4.12
executing==1.2.0
Faker==17.0.0
jedi==0.18.2
jmespath==1.0.1
kombu==5.2.4
matplotlib-inline==0.1.6
mccabe==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.37
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.14.0
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.9.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
